Apply permission check for projects and tasks

+++ backend/src/tasks/tasks.service.ts
@@ -1,9 +1,4 @@
-import {
-  Injectable,
-  InternalServerErrorException,
-  NotFoundException,
-  NotImplementedException,
-} from '@nestjs/common';
+import { Injectable, InternalServerErrorException, NotFoundException, UnauthorizedException, } from '@nestjs/common';
 import { Task } from './entities/tasks.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskDto } from './dto/task.dto';
@@ -12,24 +7,36 @@ import { UpdateTaskDto } from './dto/update-task.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Project } from '../projects/entities/projects.entity';
+import { UserDto } from '../users/dto/user.dto';
+import { PERMISSIONS } from '../permissions/permissions';
+import { UsersService } from '../users/users.service';
 
 @Injectable()
 export class TasksService {
   constructor(
     @InjectRepository(Task) private readonly taskRepository: Repository<Task>,
-    @InjectRepository(Project)
-    private readonly projectRepository: Repository<Project>,
+    @InjectRepository(Project) private readonly projectRepository: Repository<Project>,
+    private readonly usersService: UsersService,
   ) {}
 
-  async createOne(createTaskDto: CreateTaskDto): Promise<TaskDto> {
+  async createOne(
+    createTaskDto: CreateTaskDto,
+    user: UserDto,
+  ): Promise<TaskDto> {
+    const project = await this.projectRepository.findOne(
+      createTaskDto.project,
+      { relations: ['tasks', 'user'] },
+    );
+
+    if (project.user.id !== user.id && !await this.usersService.hasPermissions(user, PERMISSIONS.PROJECTS_MANAGE)) {
+      throw new UnauthorizedException(`Missing permissions for adding a task to this project`);
+    }
+
     const newTask = this.taskRepository.create({
       ...createTaskDto,
       project: null,
     });
-    newTask.project = await this.projectRepository.findOne(
-      createTaskDto.project,
-      { relations: ['tasks'] },
-    );
+    newTask.project = project;
     await this.taskRepository.save(newTask);
     newTask.project = await this.projectRepository.findOne(
       createTaskDto.project,
@@ -38,10 +45,17 @@ export class TasksService {
     return newTask;
   }
 
-  async deleteOne(filter: TasksFilter): Promise<TaskDto> {
-    const task = await this.taskRepository.findOne(filter.id);
+  async deleteOne(
+    filter: TasksFilter,
+    user: UserDto,
+  ): Promise<TaskDto> {
+    const task = await this.taskRepository.findOne(filter.id, { relations: ['project', 'project.user'] });
     if (!task) {
       throw new NotFoundException(`Could not find task with given ID!`);
+    }
+
+    if (task.project.user.id !== user.id && !await this.usersService.hasPermissions(user, PERMISSIONS.PROJECTS_MANAGE)) {
+      throw new UnauthorizedException(`Missing permissions for deleting a task of this project`);
     }
 
     const res = await this.taskRepository.delete(filter.id);
@@ -55,11 +69,17 @@ export class TasksService {
   async updateOne(
     filter: TasksFilter,
     updates: UpdateTaskDto,
+    user: UserDto,
   ): Promise<TaskDto> {
-    const task = await this.taskRepository.findOne(filter.id);
+    const task = await this.taskRepository.findOne(filter.id, { relations: ['project', 'project.user'] });
 
     if (!task) {
       throw new NotFoundException(`Could not find task with given ID!`);
+    }
+
+    const canManageProjects = await this.usersService.hasPermissions(user, PERMISSIONS.PROJECTS_MANAGE);
+    if (!((task.project.user.id === user.id && !updates.hasOwnProperty('completed')) || canManageProjects)) {
+      throw new UnauthorizedException(`Missing permissions for updating a task of this project`);
     }
 
     const updateTask = { ...task, ...updates };
